+++ htmlresources/alpermissionsetwizard/js/startup.js
@@ -1,0 +1,5 @@
+var wizard;
+
+window.onload = function() {
+    wizard = new PermissionSetWizard();
+};