+++ src/langserver/symbolsinformation/toolsGetObjectsListResponse.ts
@@ -1,0 +1,6 @@
+import { ObjectInformation } from "../../symbolsinformation/objectInformation";
+import { ToolsSymbolInformationResponse } from "./toolsSymbolInformationResponse";
+
+export class ToolsGetObjectsListResponse extends ToolsSymbolInformationResponse<ObjectInformation> {
+
+}