+++ src/langserver/symbolsinformation/toolsSymbolInformationRequest.ts
@@ -1,8 +1,10 @@
 export class ToolsSymbolInformationRequest {
     path: string | undefined;
+    includeNonAccessible: boolean;
 
-    constructor(newPath: string | undefined) {
+    constructor(newPath: string | undefined, newIncludeNonAccessible: boolean) {
         this.path = newPath;
+        this.includeNonAccessible = newIncludeNonAccessible;
     }
 
 }