+++ src/objectwizards/wizards/alPermissionSetExtensionWizardData.ts
@@ -1,0 +1,11 @@
+import { ALPermissionSetWizardData } from "./alPermissionSetWizardData";
+
+export class ALPermissionSetExtensionWizardData extends ALPermissionSetWizardData {
+    basePermissionSet: string;
+
+    constructor() {
+        super();
+        this.basePermissionSet = "";
+    }
+
+}