+++ src/symbolsinformation/permissionSetInformation.ts
@@ -1,0 +1,5 @@
+import { SymbolWithIdInformation } from "./symbolWithIdInformation";
+
+export class PermissionSetInformation extends SymbolWithIdInformation {
+
+}